Add e2e support
