+++ php_fpm/tests/conftest.py
@@ -3,6 +3,7 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import json
 import os
+from copy import deepcopy
 
 import pytest
 
@@ -13,6 +14,17 @@ from datadog_checks.php_fpm import PHPFPMCheck
 HOST = get_docker_hostname()
 HERE = os.path.dirname(os.path.abspath(__file__))
 
+DEFAULT_INSTANCE = {
+    'status_url': 'http://{}:8080/some_status'.format(HOST),
+    'ping_url': 'http://{}:8080/some_ping'.format(HOST),
+}
+
+INSTANCE_FASTCGI = {
+    'status_url': 'http://{}:9000/some_status'.format(HOST),
+    'ping_url': 'http://{}:9000/some_ping'.format(HOST),
+    'use_fastcgi': True,
+}
+
 
 @pytest.fixture
 def check():
@@ -21,19 +33,12 @@ def check():
 
 @pytest.fixture
 def instance():
-    return {
-        'status_url': 'http://{}:8080/some_status'.format(HOST),
-        'ping_url': 'http://{}:8080/some_ping'.format(HOST),
-    }
+    return deepcopy(DEFAULT_INSTANCE)
 
 
 @pytest.fixture
 def instance_fastcgi():
-    return {
-        'status_url': 'http://{}:9000/some_status'.format(HOST),
-        'ping_url': 'http://{}:9000/some_ping'.format(HOST),
-        'use_fastcgi': True,
-    }
+    return deepcopy(INSTANCE_FASTCGI)
 
 
 @pytest.fixture
@@ -47,12 +52,12 @@ def ping_url_tag_fastcgi():
 
 
 @pytest.fixture(scope='session')
-def php_fpm_instance():
+def dd_environment():
     with docker_run(
         os.path.join(HERE, 'compose', 'docker-compose.yml'),
         endpoints='http://{}:8080'.format(HOST)
     ):
-        yield
+        yield DEFAULT_INSTANCE
 
 
 @pytest.fixture
