+++ php_fpm/tox.ini
@@ -16,22 +16,11 @@ passenv =
 deps =
     -e../datadog_checks_base[deps]
     -rrequirements-dev.txt
-
-[testenv:unit]
 commands =
     pip install -r requirements.in
-    pytest -v -m "not integration"
-
-[testenv:py27-integration]
-commands =
-    pip install -r requirements.in
-    pytest -v -m "integration"
-
-[testenv:py36-integration]
-commands =
-    pip install -r requirements.in
-    pip install flup==1.0.3
-    pytest -v -m "integration"
+    unit: pytest -v -m "not integration"
+    py36-integration: pip install flup==1.0.3
+    {py27,py36}-integration: pytest -v -m "integration"
 
 [testenv:flake8]
 skip_install = true
